Changed thermistor H and L parameter scaling to match Duet 3 tool board

+++ src/Heating/Sensors/Thermistor.h
@@ -31,22 +31,22 @@ public:
 
 private:
 	// For the theory behind ADC oversampling, see http://www.atmel.com/Images/doc8003.pdf
-	static constexpr unsigned int AdcOversampleBits = 2;					// we use 2-bit oversampling
+	static constexpr unsigned int AdcOversampleBits = 2;							// we use 2-bit oversampling
 
 	void CalcDerivedParameters() noexcept;											// calculate shA and shB
 
 	// The following are configurable parameters
-	float r25, beta, shC, seriesR;											// parameters declared in the M305 command
+	float r25, beta, shC, seriesR;													// parameters declared in the M305 command
 	int8_t adcFilterChannel;
-	bool isPT1000;															// true if it is a PT1000 sensor, not a thermistor
+	bool isPT1000;																	// true if it is a PT1000 sensor, not a thermistor
 
 // Duet 3 VRef calibration doesn't work well on the MB6HC v0.6 or v1.0 so provide calibration adjustment
 #if !HAS_VREF_MONITOR || defined(DUET3)
-	int16_t adcLowOffset, adcHighOffset;
+	int8_t adcLowOffset, adcHighOffset;
 #endif
 
 	// The following are derived from the configurable parameters
-	float shA, shB;															// derived parameters
+	float shA, shB;																	// derived parameters
 
 	static constexpr int32_t OversampledAdcRange = 1u << (AdcBits + AdcOversampleBits);	// The readings we pass in should be in range 0..(AdcRange - 1)
 };
